Bumping versions of dependencies
Test vectors of encryption methods are updated. StdRng was changed from ChaCha20 to ChaCha12 in the new version of `rand`, which leads to a change in seed generation. It's okay since decryption works in a backward-compatible manner as expected. Also, it is actually expected to have a random seed in a wild, so tests just unveiled the issue which is good.

+++ src/kdf.rs
@@ -22,7 +22,7 @@ pub mod pbkdf2 {
         // Prepare an output buffer.
         let mut derived_key = [0u8; KEY_SIZE];
 
-        pbkdf2::<Hmac<Sha256>>(password, &ext_salt, iterations as usize, &mut derived_key);
+        pbkdf2::<Hmac<Sha256>>(password, &ext_salt, iterations, &mut derived_key);
 
         Zeroizing::new(derived_key)
     }
